pyproject.toml
[project]
name = "ai-chatbot-using-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents",
    "python-dotenv>=1.1.0",
]
